Move the new setting to the Door Lock tab
