+++ randovania/gui/preset_settings/generation_tab.py
@@ -35,9 +35,6 @@ class PresetGeneration(PresetTab, Ui_PresetGeneration):
         signal_handling.on_checked(self.check_major_minor, self._persist_major_minor)
         signal_handling.on_checked(self.local_first_progression_check, self._persist_local_first_progression)
 
-        # Door Lock
-        signal_handling.on_checked(self.two_sided_door_search_check, self._persist_two_sided_door_lock_search)
-
         # Logic Settings
         self.dangerous_combo.setItemData(0, LayoutLogicalResourceAction.RANDOMLY)
         self.dangerous_combo.setItemData(1, LayoutLogicalResourceAction.LAST_RESORT)
@@ -77,7 +74,6 @@ class PresetGeneration(PresetTab, Ui_PresetGeneration):
         layout = preset.configuration
 
         self.local_first_progression_check.setChecked(layout.first_progression_must_be_local)
-        self.two_sided_door_search_check.setChecked(layout.two_sided_door_lock_search)
         self.check_major_minor.setChecked(
             layout.available_locations.randomization_mode == RandomizationMode.MAJOR_MINOR_SPLIT
         )
@@ -119,9 +115,6 @@ class PresetGeneration(PresetTab, Ui_PresetGeneration):
         yield self.line_2
         yield self.experimental_generator_line
         yield self.minimal_logic_line
-        yield self.door_lock_group
-        yield self.two_sided_door_search_label
-        yield self.two_sided_door_search_check
 
     def _persist_major_minor(self, value: bool):
         mode = RandomizationMode.MAJOR_MINOR_SPLIT if value else RandomizationMode.FULL
@@ -131,10 +124,6 @@ class PresetGeneration(PresetTab, Ui_PresetGeneration):
     def _persist_local_first_progression(self, value: bool):
         with self._editor as editor:
             editor.set_configuration_field("first_progression_must_be_local", value)
-
-    def _persist_two_sided_door_lock_search(self, value: bool):
-        with self._editor as editor:
-            editor.set_configuration_field("two_sided_door_lock_search", value)
 
     def _on_dangerous_changed(self, value: LayoutLogicalResourceAction):
         with self._editor as editor:
