+++ randovania/gui/ui_files/preset_generation.ui
@@ -44,7 +44,7 @@
          <x>0</x>
          <y>0</y>
          <width>546</width>
-         <height>724</height>
+         <height>647</height>
         </rect>
        </property>
        <layout class="QVBoxLayout" name="verticalLayout_3">
@@ -113,44 +113,6 @@
             <widget class="QLabel" name="local_first_progression_label">
              <property name="text">
               <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Ensures that that your first progression is placed in your own world, ensuring you always have something to do before potentially having to wait for progression.&lt;/p&gt;&lt;p&gt;This only changes anything in multiworld sessions.&lt;/p&gt;&lt;p&gt;&lt;span style=&quot; font-weight:700;&quot;&gt;Warning:&lt;/span&gt; This will make generation fail more frequently. Use with caution in sessions with many players and random starting locations.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-             </property>
-             <property name="wordWrap">
-              <bool>true</bool>
-             </property>
-            </widget>
-           </item>
-          </layout>
-         </widget>
-        </item>
-        <item>
-         <widget class="QGroupBox" name="door_lock_group">
-          <property name="title">
-           <string>Door Lock Settings</string>
-          </property>
-          <layout class="QVBoxLayout" name="door_lock_grops">
-           <property name="leftMargin">
-            <number>9</number>
-           </property>
-           <property name="topMargin">
-            <number>9</number>
-           </property>
-           <property name="rightMargin">
-            <number>9</number>
-           </property>
-           <property name="bottomMargin">
-            <number>9</number>
-           </property>
-           <item>
-            <widget class="QCheckBox" name="two_sided_door_search_check">
-             <property name="text">
-              <string>[Experimental] Revised door solver</string>
-             </property>
-            </widget>
-           </item>
-           <item>
-            <widget class="QLabel" name="two_sided_door_search_label">
-             <property name="text">
-              <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;When using type Doors of Door Locks randomization, enables the use of a revised algorithm for deciding what doors can be.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
              </property>
              <property name="wordWrap">
               <bool>true</bool>
